fix(auth): use bigint for pollingInterval to satisfy prefer-bigint-over-int
Squawk's prefer-bigint-over-int rule fired on the integer column
type chosen to fix the type mismatch. Switch to bigint with
mode 'number' so Drizzle still maps the column to JS number
(not bigint), matching the plugin's z.number() declaration.

The 8-byte storage cost is negligible: polling intervals are
seconds, never close to 2^31. The fix is contained to
deviceCode.pollingInterval and does not affect any existing
table (the rule had not fired before because no other
table uses a numeric column).

+++ packages/database/drizzle/0001_magenta_firebrand.sql
@@ -8,7 +8,7 @@ CREATE TABLE "deviceCode" (
 	"status" text DEFAULT 'pending' NOT NULL,
 	"expires_at" timestamp NOT NULL,
 	"last_polled_at" timestamp,
-	"polling_interval" integer,
+	"polling_interval" bigint,
 	"created_at" timestamp DEFAULT now() NOT NULL,
 	"updated_at" timestamp DEFAULT now() NOT NULL,
 	CONSTRAINT "deviceCode_device_code_unique" UNIQUE("device_code"),
